% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgphobia.R
\name{pkgphobia_size}
\alias{pkgphobia_size}
\title{Get npm package install size from Packagephobia}
\usage{
pkgphobia_size(package)
}
\arguments{
\item{package}{npm package name (e.g. "express", "lodash")}
}
\value{
tibble: name (character), version (character),
publish_size (integer), install_size (integer)
}
\description{
Returns the publish and install size for an npm package.
Note: packagephobia.com uses Vercel bot protection which may
block automated requests.
}
